add collaborations and add song to playlist features

+++ src/api/playlists/handler.js
@@ -1,8 +1,12 @@
 /* eslint-disable require-jsdoc */
+const autoBind = require('auto-bind');
+
 class PlaylistsHandler {
   constructor(service, validator) {
     this._service = service;
     this._validator = validator;
+
+    autoBind(this);
   }
 
   async postPlaylistHandler(request, h) {
@@ -36,13 +40,79 @@ class PlaylistsHandler {
     };
   }
 
-  async deletePlaylistHandler(request) {
+  async deletePlaylistByIdHandler(request) {
     const {id} = request.params;
+    const {owner} = request.auth.credentials;
+
+    await this._service.verifyPlaylistOwner(id, owner);
     await this._service.deletePlaylistById(id);
 
     return {
       status: 'success',
       message: 'Playlist berhasil dihapus',
+    };
+  }
+
+  async postSongToPlaylistHandler(request) {
+    const {id: playlistId} = request.params;
+    const {songId} = request.paylaod;
+    const {owner} = request.auth.credentials;
+
+    await this._service.verifyPlaylistAccess(playlistId, owner);
+    await this._service.addSongToPlaylist(playlistId, songId, owner);
+
+    return {
+      status: 'success',
+      message: 'Lagu berhasil ditambahkan',
+    };
+  }
+
+  async getPlaylistWithSongsHandler(request) {
+    const {id: playlistId} = request.params;
+    const {id: credentialId} = request.auth.credentials;
+
+    await this._service.verifyPlaylistAccess(playlistId, credentialId);
+    const playlist = await this._service.getPlaylistById(playlistId);
+    const songs = await this._service.getSongsByPlaylistId(playlistId);
+
+    return {
+      status: 'success',
+      data: {
+        playlist: {
+          ...playlist,
+          songs,
+        },
+      },
+    };
+  }
+
+  async deleteSongFromPlaylistHandler(request) {
+    const {id: playlistId} = request.params;
+    const {songId} = request.payload;
+    const {id: credentialId} = request.auth.credentials;
+
+    await this._service.verifyPlaylistAccess(playlistId, credentialId);
+    await this._service.deleteSongFromPlaylist(
+        playlistId, songId, credentialId,
+    );
+
+    return {
+      status: 'success',
+      message: 'Lagu berhasil dihapus',
+    };
+  }
+
+  async getPlaylistActivitiesHandler(request) {
+    const {id: playlistId} = request.params;
+
+    const activities = await this._service.getPlaylistActivities(playlistId);
+
+    return {
+      status: 'success',
+      data: {
+        playlistId,
+        activities,
+      },
     };
   }
 }
